+++ src/App.scss
@@ -2512,7 +2512,7 @@ body.light-mode {
 
   @media (max-width: 672px) {
     .back {
-      background-image:linear-gradient(180deg, rgba(116, 137, 219, 0) 0%,rgba(2, 17, 104, 0.77) 100%),url('./components/images/events/rdv19.jpg');
+      background-image:linear-gradient(180deg, rgba(116, 137, 219, 0) 0%,rgba(2, 17, 104, 0.77) 100%),url('./assets/events/rdv19.jpg');
       background-position: 50% 6%;
     }
   }
@@ -5265,7 +5265,7 @@ body.dark-mode {
       background-image:linear-gradient(180deg,
         rgba(2, 17, 104, 0) 0%,
         rgba(0, 0, 0, 0.96) 100%),
-        url('./components/images/events/rdv19.jpg');
+        url('./assets/events/rdv19.jpg');
       background-position: 50% 0%;
     }
   }
@@ -5275,7 +5275,7 @@ body.dark-mode {
       background-image:linear-gradient(180deg,
         rgba(2, 17, 104, 0) 0%,
         rgba(0, 0, 0, 0.96) 100%),
-        url('./components/images/events/rdv19.jpg');
+        url('./assets/events/rdv19.jpg');
       background-position: 50% 6%;
     }
   }
