+++ src/assets/hostels/girnar.jpeg
