+++ src/assets/hostels/himadri.png
