+++ src/assets/hostels/kailash.png
