+++ src/assets/hostels/shivalik.png
