+++ src/assets/hostels/udaigiri.jpg
