+++ src/assets/hostels/vindy.jpg
