+++ src/components/home.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import * as Icon from 'react-feather';
 import { useWindowSize } from 'react-use';
-import iitd from './images/home/iitd.jpg';
-import iitSign from './images/home/iit-sign.jpg';
-import graffiti from './images/home/graffiti.png';
-import dogra from './images/home/dogra-roof.jpg';
-import lhc from './images/home/lhc.jpg';
+import iitd from '../assets/home/iitd.jpg';
+import iitSign from '../assets/home/iit-sign.jpg';
+import graffiti from '../assets/home/graffiti.png';
+import dogra from '../assets/home/dogra-roof.jpg';
+import lhc from '../assets/home/lhc.jpg';
 
 function Home() {
   const windowSize = useWindowSize();
