+++ src/data/events.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import rdv from '../images/events/rdv.svg';
-import tryst from '../images/events/tryst.svg';
-import convergenz from '../images/events/convergenz.svg';
-import literati from '../images/events/literati.svg';
-import sportech from '../images/events/sportech.webp';
-import BECon from '../images/events/BECon.png';
-import rdvBg from '../images/events/rdv19.jpg';
-import trystBg from '../images/events/trystBg.jpeg';
-import convergenzBg from '../images/events/convergenzBG.jpg';
-import literatiBg from '../images/events/literatiBG.png';
-import BEConBg from '../images/events/BEConBG.jpg';
-import sportechBg from '../images/events/sportechBG.jpg';
+import rdv from '../assets/events/rdv.svg';
+import tryst from '../assets/events/tryst.svg';
+import convergenz from '../assets/events/convergenz.svg';
+import literati from '../assets/events/literati.svg';
+import sportech from '../assets/events/sportech.webp';
+import BECon from '../assets/events/BECon.png';
+import rdvBg from '../assets/events/rdv19.jpg';
+import trystBg from '../assets/events/trystBg.jpeg';
+import convergenzBg from '../assets/events/convergenzBG.jpg';
+import literatiBg from '../assets/events/literatiBG.png';
+import BEConBg from '../assets/events/BEConBG.jpg';
+import sportechBg from '../assets/events/sportechBG.jpg';
 
 const events = [
   {
